[backends] Drop drizzle-orm from 21-elysia fixture to clear CVE
The fixture pinned drizzle-orm@0.40.1, which Socket flagged for high-severity
CVE-2026-39356 (GHSA-gpj5-g38j-94v9, SQL injection via improperly escaped
identifiers; fixed in 0.45.2). drizzle was only the vehicle that pulled in
@planetscale/database; the test actually verifies NFT resolves the package's
CJS conditional export (dist/cjs/index.js). Rewrote @repro/data to use the
@planetscale/database Client/Connection API directly, removing drizzle-orm,
its pnpm patch, and the related lockfile entries. The files.json assertion and
the build still pass.
