+++ packages/backends/test/fixtures/21-elysia-vercel-repro/packages/data/src/database/database.ts
@@ -1,30 +1,9 @@
-import { AsyncLocalStorage } from 'async_hooks';
-import { createDrizzleClient } from './drizzleClient';
-import type { DrizzleDatabase, DrizzleDatabaseTransaction } from './drizzleClient';
+import { createDatabaseClient } from './client';
+import type { DatabaseClient } from './client';
 
 // Top-level instantiation at module-load time, mirroring azzapp's pattern.
 // This forces @planetscale/database (transitive workspace dep) to be
 // resolvable at the moment data's index.cjs is loaded.
-const drizzleClient = createDrizzleClient();
+const client = createDatabaseClient();
 
-const transactionStorage = new AsyncLocalStorage<DrizzleDatabaseTransaction>();
-const usePrimaryStorage = new AsyncLocalStorage<boolean>();
-
-export const db = (): DrizzleDatabase => {
-  return transactionStorage.getStore() ?? drizzleClient;
-};
-
-export const runWithPrimary = async <T>(
-  callback: () => Promise<T>,
-): Promise<T> => usePrimaryStorage.run(true, callback);
-
-export const transaction = async <T>(
-  callback: (tx: { rollback(): void }) => Promise<T>,
-): Promise<T> =>
-  db().transaction(tx => {
-    return new Promise<T>((resolve, reject) => {
-      transactionStorage.run(tx, () => {
-        callback(tx).then(resolve, reject);
-      });
-    });
-  });
+export const db = (): DatabaseClient => client;
