+++ packages/backends/test/fixtures/21-elysia-vercel-repro/packages/data/src/index.ts
@@ -1,4 +1,4 @@
-export * from './database/drizzleClient';
+export * from './database/client';
 export * from './database/database';
 export * from './helpers/createId';
 export * from './queries';
