+++ packages/backends/test/fixtures/21-elysia-vercel-repro/pnpm-workspace.yaml
@@ -5,12 +5,10 @@ allowBuilds:
   unrs-resolver: true
 nodeLinker: isolated
 publicHoistPattern:
-  # The crux of the repro: drizzle-orm and @adobe/fetch hoist to the root, but
-  # @planetscale/database does NOT (it's a peerDep of the patched drizzle-orm).
-  # That asymmetry is what triggers the missing-dependency bug being tested.
+  # The crux of the repro: @adobe/fetch hoists to the root, but
+  # @planetscale/database does NOT. That asymmetry is what triggers the
+  # missing-dependency bug being tested — the CJS variant of
+  # @planetscale/database lives in a non-hoisted, nested node_modules.
   - '@planetscale/database'
   - '@adobe/fetch'
-  - 'drizzle-orm'
 strictPeerDependencies: false
-patchedDependencies:
-  drizzle-orm@0.40.1: patches/drizzle-orm@0.40.1.patch
